requirement.txt
streamlit
langchain
langchain_community
langchain_core
langchain_openai
trafilatura
elevenlabs
python-dotenv
openai>=1.0.0
